refactor: move reused functions to funtil module
Fun utility functions! Funtils! I'll see myself out..
