+++ src/funtils.js
@@ -1,0 +1,5 @@
+// string -> number
+export const toDecimal = (str) => parseInt(str, 10);
+
+// number -> number
+export const sum = (a, b) => a + b;
